# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Monarr API
  version: "1"
  description: >
    Native Monarr API (spec-first; blueprint §7). Served under /api/v1.
    This is the only API the web UI uses. Phase 0 surface: system status,
    health, scheduler tasks, and the SSE event stream.
  license:
    name: GPL-3.0
paths:
...
  /indexers:
    get:
      operationId: listIndexers
      summary: List indexers
      responses:
        "200":
          description: Indexers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Indexer"
    post:
      operationId: addIndexer
      summary: Add an indexer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexerInput"
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indexer"
        "400":
          description: Invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /indexers/test:
    post:
      operationId: testIndexer
      summary: Test an indexer config (t=caps)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexerInput"
      responses:
        "200":
          description: Reachable.
        "400":
          description: Failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /indexers/{id}:
    delete:
      operationId: deleteIndexer
      summary: Remove an indexer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Removed. }
  /downloadclients:
    get:
      operationId: listDownloadClients
      summary: List download clients
      responses:
        "200":
          description: Clients.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DownloadClientConfig"
    post:
      operationId: addDownloadClient
      summary: Add a download client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadClientInput"
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadClientConfig"
        "400":
          description: Invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /downloadclients/test:
    post:
      operationId: testDownloadClient
      summary: Test a download client config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadClientInput"
      responses:
        "200": { description: Reachable. }
        "400":
          description: Failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /downloadclients/{id}:
    delete:
      operationId: deleteDownloadClient
      summary: Remove a download client
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Removed. }
  /library/{id}/releases:
    get:
      operationId: searchReleases
      summary: Interactive search for a movie, episode, or season pack
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: season
          in: query
          required: false
          schema: { type: integer }
        - name: episode
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Ranked candidates with rejection reasons attached.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReleaseCandidate"
        "404":
          description: Unknown item/season/episode.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: No enabled indexers.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /grab:
    post:
      operationId: grabRelease
      summary: Send a release to a download client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrabRequest"
      responses:
        "201":
          description: Queued; returns the download row id.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: integer, format: int64 }
        "400":
          description: Invalid or no client for protocol.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /queue/{id}:
    delete:
      operationId: removeQueueItem
      summary: Remove a queue row (optionally from the client too)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: fromClient
          in: query
          required: false
          schema: { type: boolean }
      responses:
        "204": { description: Removed. }
        "404":
          description: Unknown row.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /customformats:
    get:
      operationId: listCustomFormats
      summary: Custom-format scoring rules
      responses:
        "200":
          description: Rules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomFormat"
    post:
      operationId: addCustomFormat
      summary: Add a scoring rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomFormatInput"
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomFormat"
        "400":
          description: Invalid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /customformats/{id}:
    delete:
      operationId: deleteCustomFormat
      summary: Remove a scoring rule
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Removed. }
  /importlists:
    get:
      operationId: listImportLists
      summary: Import list sources
      responses:
        "200":
          description: Lists.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImportList"
    post:
      operationId: addImportList
      summary: Add an import list source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportListInput"
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportList"
        "400":
          description: Invalid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /importlists/{id}:
    delete:
      operationId: deleteImportList
      summary: Remove an import list source
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Removed. }
  /library/bulk:
    post:
      operationId: bulkEditLibrary
      summary: Apply monitoring/profile changes to many items at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: integer, format: int64 }
                monitored:
                  type: boolean
                qualityProfileId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Applied.
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated:
                    type: integer
...
  /notifiers:
    get:
      operationId: listNotifiers
      summary: List notification targets
      responses:
        "200":
          description: Notifiers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notifier"
    post:
      operationId: addNotifier
      summary: Add a notification target
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifierInput"
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notifier"
        "400":
          description: Invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /notifiers/test:
    post:
      operationId: testNotifier
      summary: Send a test notification with an unsaved config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifierInput"
      responses:
        "200":
          description: Delivered.
        "400":
          description: Failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /notifiers/{id}:
    delete:
      operationId: deleteNotifier
      summary: Remove a notification target
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Removed. }
        "404":
          description: Unknown notifier.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /calendar:
    get:
      operationId: getCalendar
      summary: Episodes airing and movies/books released in a date range
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string }
        - name: end
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Calendar entries, date-ordered.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEntry"
...
  /blocklist/{id}:
    delete:
      operationId: removeBlocklistEntry
      summary: Unblock a release
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Removed. }
...
components:
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    IndexerInput:
      type: object
      required: [name, url, protocol]
      properties:
        name: { type: string }
        url: { type: string }
        apiKey: { type: string }
        protocol: { type: string, enum: [torrent, usenet] }
        categories:
          type: array
          items: { type: integer }
        enabled: { type: boolean, default: true }
    Indexer:
      allOf:
        - $ref: "#/components/schemas/IndexerInput"
        - type: object
          required: [id]
          properties:
            id: { type: integer, format: int64 }
    DownloadClientInput:
      type: object
      required: [type, name, url]
      properties:
        type: { type: string, enum: [qbittorrent, sabnzbd, transmission, deluge, nzbget] }
        name: { type: string }
        url: { type: string }
        username: { type: string }
        password: { type: string }
        category: { type: string, default: monarr }
        enabled: { type: boolean, default: true }
    DownloadClientConfig:
      allOf:
        - $ref: "#/components/schemas/DownloadClientInput"
        - type: object
          required: [id]
          properties:
            id: { type: integer, format: int64 }
    Rejection:
      type: object
      required: [code, reason]
      properties:
        code: { type: string }
        reason: { type: string }
    ReleaseCandidate:
      type: object
      required: [title, downloadUrl, indexer, protocol, size, seeders, age, quality, score, accepted, isUpgrade, rejections]
      properties:
        title: { type: string }
        downloadUrl: { type: string }
        infoUrl: { type: string }
        indexer: { type: string }
        protocol: { type: string }
        size: { type: integer, format: int64 }
        seeders: { type: integer }
        age: { type: string }
        quality: { type: string }
        score:
          type: integer
          description: Custom-format score sum (Phase 5).
        formats:
          type: array
          items: { type: string }
          description: Names of matched custom formats.
        accepted: { type: boolean }
        isUpgrade: { type: boolean }
        rejections:
          type: array
          items: { $ref: "#/components/schemas/Rejection" }
    GrabRequest:
      type: object
      required: [mediaItemId, title, downloadUrl, protocol]
      properties:
        mediaItemId: { type: integer, format: int64 }
        season: { type: integer }
        episode: { type: integer }
        title: { type: string }
        downloadUrl: { type: string }
        indexer: { type: string }
        protocol: { type: string }
        size: { type: integer, format: int64 }
    CustomFormatInput:
      type: object
      required: [name, pattern]
      properties:
        name: { type: string }
        pattern:
          type: string
          description: Case-insensitive RE2 matched against release titles.
        score:
          type: integer
          default: 0
    CustomFormat:
      allOf:
        - $ref: "#/components/schemas/CustomFormatInput"
        - type: object
          required: [id]
          properties:
            id: { type: integer, format: int64 }
    ImportListInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type:
          type: string
          enum: [tmdb-popular, tmdb-top, trakt-list]
        config:
          type: object
          additionalProperties: { type: string }
          description: "trakt-list: {user, slug, clientId}."
        kind:
          $ref: "#/components/schemas/MediaKind"
        rootFolderId: { type: integer, format: int64 }
        qualityProfileId: { type: integer, format: int64 }
        monitored: { type: boolean, default: true }
        enabled: { type: boolean, default: true }
    ImportList:
      allOf:
        - $ref: "#/components/schemas/ImportListInput"
        - type: object
          required: [id]
          properties:
            id: { type: integer, format: int64 }
...